+++ NEWS.md
@@ -8,6 +8,7 @@
 
 * Fixed the bug of `summarise.tbl_ts()` and `tsummarise.tbl_ts()`, when calling functions with no parameters like `dplyr::n()`. (@mitchelloharawild)
 * One grouping level should be dropped for the consistency with `dplyr::summarise()` for a grouped data frame. (@mitchelloharawild)
+* Fixed incorrect group and key indices.
 
 # tsibble 0.1.0